indiaml/pyproject.toml
[project]
name = "indiaml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.8.2",
    "exa-py>=1.14.9",
    "ipykernel>=6.29.5",
    "openai>=1.66.2",
    "openreview-py>=1.46.3",
    "orjson>=3.10.18",
    "pandas>=2.2.3",
    "pycountry>=24.6.1",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.9.1",
    "pyfunctional>=1.5.0",
    "pymupdf4llm>=0.0.17",
    "python-dotenv>=1.0.1",
    "rapidfuzz>=3.13.0",
    "sqlalchemy>=2.0.39",
]
